Refactor UI graph canvas
Related to #1

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/augustodia/palestra-grafos-uninter/issues/1?shareId=XXXX-XXXX-XXXX-XXXX).
